pyproject.toml
[tool.poetry]
name = "elu-twins"
version = "0.1.0"
description = "A set to tools to emulate devices related with electro mobility"
authors = ["Alfonso Sastre <alfonso@elumobility.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
sqlmodel = "^0.0.16"
ocpp = "^0.26.0"
pandas = "^2.2.2"
celery = {extras = ["redis"], version = "^5.4.0"}
flower = "^2.0.1"
asgiref = "^3.8.1"
websockets = "^12.0"
aiohttp = "^3.9.5"
numpy = "^1.26.4"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
loguru = "^0.7.2"
bcrypt = "^4.1.2"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
mysql-connector-python = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
